+++ src/day02/p2.rs
@@ -4,7 +4,7 @@ struct Loc {
     aim: i32,
 }
 
-pub fn solve(inp: Vec<&str>) -> Result<i32, Box<dyn std::error::Error>> {
+pub fn solve(inp: Vec<&str>) -> Result<i64, Box<dyn std::error::Error>> {
     let mut my_loc = Loc {
         depth: 0,
         line: 0,
